ハッシュタグレンダリングの関数を切り出して共通化
